ui: make route to component mapping explicit
After this commit instead of the App component assuming the convention
that the component associated with the route expects the whole GraphQL
response in its "data" property the mapping is moved to mini render
functions explicitly defined for each route.

This will allow to transition currently non-generic components (i.e ones
that their fragment is based on the root Query) to generic ones.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>